Inital new project
The project is supposed to answer the question whether solutions
for a cube of sudokus do exist.

We need a general and reasonably efficient sudoku solver for that.
This is what the Sudoku class aims for. Currently it only properly
initializes the columns, rows and boxes in a general sudoku, i.e.
not a 9 x 9 square but a general (n * m) x (n * m) square.