Refactor Cell class and import
The cell class now knows more about itself and the sudoku it is in
and uses properties for automation:

- Its row, col, and parent attributes are readonly and show its
  row and column number and allow access to the parent sudoku.
- Its value is the `val` property, that can be set, thus solving
  the cell, instead of using different properties.
- To exclude a candidate from a cell use xclude, that autmatically
  sets the value if only a single candidate remains.
- A cell can be copied (for backups on the stack) an printed, the
  latter returns a number of lines containing ascii art for the
  cell content, that either shows the remaininig candidates or
  the solved number. Thus printing is supposed to print intermediate
  states as well for visual feedback.

There is a new rule implementation: apply_singlepos, that - given
a number - checks in all rows, columns and boxes if only a single
position remains for the number and sets it there as well.

Finally import is cleaned up: There is an Import class that reads
a file, creates the corresponding Sudoko instance and the givens.
The Import class is now independent of the Sudoku class.