Add backup and restore methods
A cell's backup is just the cell's value that is returned, with
the added trick to return a copy if the value is mutable. This
returns the full state of a cell.

A sudoku's backup is just the list of all cell backups, and is
saved on a backup stack.