Implement trying
Add another (read-only) property name to Cell, to print the cell in
error and log messages.

Do not automatically set the value when exclusion lets the
candidate number drop to 1: Whenever we set a value we need to
immediately exclude it from all other cells in all houses it is
in, otherwise we might set two cells in a house to the same value.

This automatism must be placed in the Sudoku class, so add the new
setcell method.

When the last value is attempted to be removed from a cell we can
indeed raise an Unsolvable exception, Together with setcell,  this
probably detects false tries quicker.

Add a backup class, an object because we need to save the cell
values and the current number of remaining (i.e. not solved) cells,
that serve as a progress indicator as well as a fast completed
condition.

The type of house (row, col, box) gets a name now as well, for better
error and log messages.

Givens are now just remembered during `add_given` and then applied
first using `apply_givens` as this may already raise an Unsolvable
exception.

Solve methods and test program had to be rewritten almost completely.