Make sudoku a special case of magic square
We start with NCell derived from BaseCell that is a general cell
that can hold one out of N symbols. All methods of class Cell except
print can then be moved to NCell. The backup method is renamed to
state.

Now all methods that are valid for magic squares as well are moved
there. To allow for more generality, Magicsquare has three new
properties: houses, housenames and myhouse, all of them lists
having two entries.

The houses list contains lists holding the different house types:
rows and columns. The housenames has a string name for each house
type and myhouse contains a function for each housetype that can
be given a cell and returns the house that has the cell as the member.

So now Sudoku becomes a Magicsquare that has another house type (box)
and Magicsquare has the two simple strategies: singlecandidate
and singlepos.