Implement pairs in Euleros
We hold all possible pairs as keys in a dict. The value is
initialized to the set of all cells. Whenever a cell can be
excluded from holding a certain pair, the cell is removed from
the set. When a cell becomes knowns to hold a fixed pair, the set
is replaced by the cell itself.