Add read and write support to euleros
We read and write a single cell given as a string of zero or more
uppercase characters followed by zero or more decimal digits. Empty
cells are everything that does match that pattern, we currently
use a single underscore. The values must be whitespace separated,
in as many rows as the puzzle has.

Other corrections:

- fix the value confusion: values are in the 1..n range
- do not import cell itself as the name might be shadowed by a
  local variable of the same name
- house must be iterated over the myhouse property