Change how a cell gets a value
We now put more logic into a single cell: Instead of a setter for
the value, we must call the setval method, where a comment can be
supplied as well. The cell does the logging and at the end calls
the parent's cellgotval method if that exists.

The Magicsquare's cellgotval then replaces the former setcell
method, except for the logging of course, and will in turn try to
call a parent's cellgotval method if a parent with such a method
exists. This can later be used for Euleros to react on a cell
in one of its magic squares to receive a value.

Finally we create a BasePuzzle class where we now locate the
solve, backup and restore methods which are all generic.
The Magicsquare restorestate method now only restores the
Magicsquare specific part of the state.

Also drop the unused givens property from Magicsquare