Adapt eulero to new cell value setting function
We rename the function setcell to cellgotvalue and thus make sure it
will called whenever a cell in any of the two magic squares changes
from a set to a fixed value. That nicely solves the problem how to
reliably intercept all value settings.

The two magic squares of an eulero only need to obtain the parent
and pos properties, and the cellgotval function first needs to
accesss the square value to obtain the position (it could of
course compare or look it up in a dict)

The new rule_singlepairpos is the only eulero-specific rule to
detect cell values: When there is only one remaining position for
a pair, then both cells at that position must have the respective
value.

Other changes:

- To obtain a Magicsquare cell at a given position, use the `getcell`
  method, the `cell` method does not exist.
- initialize the remain property of the eulero and decrement it in
  cellgotval. We could also use property instead. but we need the
  remain property for the generic solver.
- Add a test-program in a if __name__ == '__main__' block