More eulero fixes
Drop the Eulorosquare class, now that we found a better solution
for the value receiving notification problem.

Eulero needs to be a subclass of BasePuzzle

Enumerate with an iterator that yields pairs values must use a
nested pair as well.

Fix row and col numbers in debug message (these are 1 based)

As a quick fix use quickprint for the required print method