Propagate exclusion of cell values
When a value was excluded from a cell propagate that upwards to
the parent in the same way as we do when a cell receives a value.
Note that a cell that receives a value will also create some
exclude notifications, if its set of possible values had more than
one element.

The BasePuzzle class now adds debug messages on save/restore, and
keeps track on the maximum level when values are tried.
When all values have been tried and no solution was found, do not
raise an Unsolvable error, just log that no solution was found

Drop the frequent "Try singlepos rule" and Found single candidate,
we are now convinced that the rules work the right way, and the
comments are good enough to see the progress.