Fix eulero solver
The hardest part was figuring out that a shallow copy of the pairs
dict is too shallow: the new dict references the same values. We
must copy the values as well.

For better readable debug messages we add pairname and posname
methods.

We now use the notification methods cellgotval and cellnotval,
that are much simpler than the previous setpair method.

Also log to stdout, which allows us to see the print() output at
the right place.

The solver now successfully solves the eulero test cases.